refactor: rename header client_services -> client_service
